backend/requirements.txt
django==5.0.6
djangorestframework==3.15.1
psycopg[binary]==3.1.19
Pillow==10.3.0
